update `ansi_colors.py` to `styles.py`, and other small changes
