+++ operate/utils/prompt_util.py
@@ -1,5 +1,5 @@
 from prompt_toolkit.styles import Style as PromptStyle
-from operate.prompts.prompt import (
+from operate.prompts.prompts import (
     VISION_PROMPT,
     ACCURATE_PIXEL_COUNT,
     ACCURATE_MODE_VISION_PROMPT,
